requirements.txt
colorama==0.4.6
dotenv==0.9.9
iniconfig==2.0.0
mysql-connector-python==9.2.0
packaging==24.2
pluggy==1.5.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pytest==8.3.4
pytest-mock==3.14.0
python-dotenv==1.0.1
setuptools==75.8.0
